Esquema pra passar no ABC, achei um erro Comecando terceiro metodo Adicionando contagem de nodos removidos com certo depth A3 estranho, removido constantes do graph
Adições:
-depth calcul.
-A4
-Propagação de bits em vetor feito.
-propagateandeletev2 virou titular
-apply image AUX struct REFEITO (continua mais lento que recursivo)
-Cut AIG adicionado
-Remocao de nodos REFEITA.
-adicionado varias execucoes na mesma chamada.

Bug fix: 
-Binario para decimal arrumado
-ordem correta dos inputs dos nodos AND
-adicionado remocao de POs (arrumado)
-applymnist bombando

+++ graph.cpp
@@ -1897,7 +1897,7 @@ void graph::propagateAndDeleteAll(mnist& mnist_obj,int option,float min_th,int a
 //            smallest=*min_element(begin(depth_counter),end(depth_counter));
 //            cout<<biggest<<">>>>>>>>>>"<<smallest<<endl;
             for(int k=0;k<new_ths.size();k++)
-                new_ths[k]=((1-min_th)*((-depth_counter[k]/(biggest-1))+(biggest/(biggest-1))))+min_th;
+                new_ths[k]=((1-min_th)*((((float)-depth_counter[k])/(biggest-1))+((float)biggest/(biggest-1))))+min_th;
         }
     }
     else
